Move find_path_from_directory to shared module_loading library
