+++ shared/module_loading/pyproject.toml
@@ -25,6 +25,7 @@ classifiers = [
 
 dependencies = [
     'importlib_metadata>=6.5;python_version<"3.12"',
+    "pathspec>=0.9.0",
 ]
 
 [dependency-groups]
