+++ task-sdk/pyproject.toml
@@ -79,6 +79,7 @@ dependencies = [
     # End of shared configuration dependencies
     # Start of shared module-loading dependencies
     'importlib_metadata>=6.5;python_version<"3.12"',
+    "pathspec>=0.9.0",
     # End of shared module-loading dependencies
 ]
